+++ pages/_error.tsx
@@ -2,7 +2,7 @@ import NotFound from './404'
 import Bugsnag from '@hashicorp/platform-runtime-error-monitoring'
 
 function Error({ statusCode }) {
-  return <NotFound statusCode={statusCode} />
+  return <NotFound />
 }
 
 Error.getInitialProps = ({ res, err }) => {
