--- pages/docs/[[...page]].jsx
@@ -1,49 +1,0 @@
-import { productName, productSlug } from 'data/metadata'
-import DocsPage from '@hashicorp/react-docs-page'
-import ProviderTable from 'components/provider-table'
-// Imports below are only used server-side
-import {
-  generateStaticPaths,
-  generateStaticProps,
-} from '@hashicorp/react-docs-page/server'
-
-//  Configure the docs path
-const BASE_ROUTE = 'docs'
-const NAV_DATA = 'data/docs-nav-data.json'
-const CONTENT_DIR = 'content/docs'
-const PRODUCT = { name: productName, slug: 'terraform-website' }
-
-function DocsLayout(props) {
-  return (
-    <DocsPage
-      additionalComponents={{ ProviderTable }}
-      baseRoute={BASE_ROUTE}
-      product={PRODUCT}
-      staticProps={props}
-    />
-  )
-}
-
-export async function getStaticPaths() {
-  const paths = await generateStaticPaths({
-    navDataFile: NAV_DATA,
-    localContentDir: CONTENT_DIR,
-    product: PRODUCT,
-  })
-  return { paths, fallback: false }
-}
-
-export async function getStaticProps({ params }) {
-  const props = await generateStaticProps({
-    navDataFile: NAV_DATA,
-    localContentDir: CONTENT_DIR,
-    params,
-    product: PRODUCT,
-    githubFileUrl(path) {
-      return `https://github.com/hashicorp/${PRODUCT.slug}/blob/master/${path}`
-    },
-  })
-  return { props }
-}
-
-export default DocsLayout
