+++ pages/docs/[[...page]].tsx
@@ -1,0 +1,38 @@
+import { productName, productSlug } from 'data/metadata'
+import DocsPage from '@hashicorp/react-docs-page'
+import ProviderTable from 'components/provider-table'
+// Imports below are only used server-side
+import { getStaticGenerationFunctions } from '@hashicorp/react-docs-page/server'
+
+//  Configure the docs path
+const BASE_ROUTE = 'docs'
+const NAV_DATA = 'data/docs-nav-data.json'
+const CONTENT_DIR = 'content/docs'
+const PRODUCT = { name: productName, slug: 'terraform' } as const
+
+const SOURCE_REPO = 'terraform-website'
+const DEFAULT_BRANCH = 'master'
+
+function DocsLayout(props) {
+  return (
+    <DocsPage
+      additionalComponents={{ ProviderTable }}
+      baseRoute={BASE_ROUTE}
+      product={PRODUCT}
+      staticProps={props}
+    />
+  )
+}
+
+const { getStaticPaths, getStaticProps } = getStaticGenerationFunctions({
+  strategy: 'fs',
+  localContentDir: CONTENT_DIR,
+  navDataFile: NAV_DATA,
+  product: SOURCE_REPO,
+  githubFileUrl(filepath) {
+    return `https://github.com/hashicorp/${SOURCE_REPO}/blob/${DEFAULT_BRANCH}/${filepath}`
+  },
+})
+export { getStaticPaths, getStaticProps }
+
+export default DocsLayout
