+++ pages/guides/[[...page]].tsx
@@ -1,23 +1,28 @@
 import { productName, productSlug } from 'data/metadata'
 import DocsPage from '@hashicorp/react-docs-page'
-import otherDocsData from 'data/other-docs-nav-data.json'
+import ProviderTable from 'components/provider-table'
 // Imports below are only used server-side
 import { getStaticGenerationFunctions } from '@hashicorp/react-docs-page/server'
 import path from 'path'
 import { remarkRewriteAssets } from 'lib/remark-rewrite-assets'
 
 //  Configure the docs path
-const BASE_ROUTE = 'configuration'
+const BASE_ROUTE = 'guides'
 const NAV_DATA = path.join(
   process.env.NAV_DATA_DIRNAME,
   BASE_ROUTE + '-nav-data.json'
 )
 const CONTENT_DIR = path.join(process.env.CONTENT_DIRNAME, BASE_ROUTE)
-const PRODUCT = { name: productName, slug: productSlug }
+const PRODUCT = { name: productName, slug: productSlug } as const
 
-export default function ConfigurationLayout(props) {
+export default function GuidesLayout(props) {
   return (
-    <DocsPage baseRoute={BASE_ROUTE} product={PRODUCT} staticProps={props} />
+    <DocsPage
+      additionalComponents={{ ProviderTable }}
+      baseRoute={BASE_ROUTE}
+      product={PRODUCT}
+      staticProps={props}
+    />
   )
 }
 
@@ -26,7 +31,6 @@ const { getStaticPaths, getStaticProps } = getStaticGenerationFunctions(
     process.env.PREVIEW_FROM_REPO === 'terraform'
     ? {
         strategy: 'fs',
-        basePath: BASE_ROUTE,
         localContentDir: CONTENT_DIR,
         navDataFile: NAV_DATA,
         product: PRODUCT.slug,
