+++ pages/logos/index.tsx
