+++ pages/registry/[[...page]].tsx
@@ -3,16 +3,16 @@ import DocsPage from '@hashicorp/react-docs-page'
 import ProviderTable from 'components/provider-table'
 import otherDocsData from 'data/other-docs-nav-data.json'
 // Imports below are only used server-side
-import {
-  generateStaticPaths,
-  generateStaticProps,
-} from '@hashicorp/react-docs-page/server'
+import { getStaticGenerationFunctions } from '@hashicorp/react-docs-page/server'
 
 //  Configure the path
 const BASE_ROUTE = 'registry'
 const NAV_DATA = 'data/registry-nav-data.json'
 const CONTENT_DIR = 'content/registry'
-const PRODUCT = { name: productName, slug: 'terraform-website' }
+const PRODUCT = { name: productName, slug: 'terraform' } as const
+
+const SOURCE_REPO = 'terraform-website'
+const DEFAULT_BRANCH = 'master'
 
 function RegistryLayout(props) {
   // add the "other docs" section to the bottom of the nav data
@@ -29,26 +29,15 @@ function RegistryLayout(props) {
   )
 }
 
-export async function getStaticPaths() {
-  const paths = await generateStaticPaths({
-    navDataFile: NAV_DATA,
-    localContentDir: CONTENT_DIR,
-    product: PRODUCT,
-  })
-  return { paths, fallback: false }
-}
-
-export async function getStaticProps({ params }) {
-  const props = await generateStaticProps({
-    navDataFile: NAV_DATA,
-    localContentDir: CONTENT_DIR,
-    params,
-    product: PRODUCT,
-    githubFileUrl(path) {
-      return `https://github.com/hashicorp/${PRODUCT.slug}/blob/master/${path}`
-    },
-  })
-  return { props }
-}
+const { getStaticPaths, getStaticProps } = getStaticGenerationFunctions({
+  strategy: 'fs',
+  localContentDir: CONTENT_DIR,
+  navDataFile: NAV_DATA,
+  product: SOURCE_REPO,
+  githubFileUrl(filepath) {
+    return `https://github.com/hashicorp/${SOURCE_REPO}/blob/${DEFAULT_BRANCH}/${filepath}`
+  },
+})
+export { getStaticPaths, getStaticProps }
 
 export default RegistryLayout
